Time: 46 ms (54.42%), Space: 65.1 MB (50.79%) - LeetHub

+++ 2300-successful-pairs-of-spells-and-potions/2300-successful-pairs-of-spells-and-potions.java
@@ -24,7 +24,7 @@ int low=0;
 int high=m-1;
 while(low<=high){
 int mid=low+(high-low)/2;
-long product=(long)(potions[mid]*spells[i]);
+long product=((long)potions[mid]*(long)spells[i]);
 if(product>=success){
 high=mid-1;}
 else {
